Support uint16 and uint32 in to_torch (#36660)
### Ticket

https://github.com/tenstorrent/tt-metal/issues/31150

### Problem description

This PR was created to facilitate merging the following PR from the
fork: https://github.com/tenstorrent/tt-metal/pull/32350

### What's changed

- Remove nanobind code that explicitly converted the tensor to the
signed variant. This changes behavior for both torch and numpy results.
- Update tests that relied on the torch/numpy tensor to have signed
values as a result of the `to_torch` conversion. Use the
`convert_to_signed_tensor` function to match the previous behavior.

### Checklist

- [x] [![All post-commit
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml/badge.svg?branch=mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml?query=branch:mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)
- [x] [![Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml/badge.svg?branch=mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml?query=branch:mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)
- [x]
[![cpp-unit-tests](https://github.com/tenstorrent/tt-metal/actions/workflows/tt-metal-l2-nightly.yaml/badge.svg?branch=mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)](https://github.com/tenstorrent/tt-metal/actions/workflows/tt-metal-l2-nightly.yaml?query=branch:mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)
- [ ] New/Existing tests provide coverage for changes


#### Model tests

If your changes cover model-related code, you should run tests
corresponding to affected models and platforms (Single card, T3K,
Galaxy). "Choose your pipeline" workflows facilitate running multiple
kinds of tests in a single run. Each offers `models-mandatory` and
`models-extended` presets.
The former includes a minimal set of tests, to be run always. The latter
extends that with additional ones - use your best judgement in deciding
which is the most appropriate for your PR.

- [ ] [![(Single) Choose your
pipeline](https://github.com/tenstorrent/tt-metal/actions/workflows/pipeline-select.yaml/badge.svg?branch=mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)](https://github.com/tenstorrent/tt-metal/actions/workflows/pipeline-select.yaml?query=branch:mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)
- [x] `models-mandatory` preset (runs: [Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
and [Frequent model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/fast-dispatch-full-regressions-and-models.yaml))
- [ ] `models-extended` preset (runs: the mandatory tests, plus
[Demo](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml)
and [Model
perf](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
tests)
  - [ ] other selection - specify runs

- [ ] [![(T3K) Choose your
pipeline](https://github.com/tenstorrent/tt-metal/actions/workflows/pipeline-select-t3k.yaml/badge.svg?branch=mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)](https://github.com/tenstorrent/tt-metal/actions/workflows/pipeline-select-t3k.yaml?query=branch:mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)
- [ ] `models-mandatory` preset (runs: [Unit
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml))
- [ ] `models-extended` preset (runs: the mandatory tests, plus
[Demo](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml)
and [Model
perf](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-model-perf-tests.yaml)
tests)
  - [ ] other selection - specify runs

- [ ] [![(Galaxy) Choose your
pipeline](https://github.com/tenstorrent/tt-metal/actions/workflows/pipeline-select-galaxy.yaml/badge.svg?branch=mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)](https://github.com/tenstorrent/tt-metal/actions/workflows/pipeline-select-galaxy.yaml?query=branch:mirror/MaximArtemovEPAM/maxim_artemov/TTNN-31150-uint16-uint32-to_torch)
- [ ] `models-mandatory` preset (runs: [Quick
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-quick.yaml))
- [ ] `models-extended` preset (runs: the mandatory tests, plus
[Demo](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-demo-tests.yaml)
and [Model
perf](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-model-perf-tests.yaml)
tests)
  - [ ] other selection - specify runs

---------

+++ tests/ttnn/unit_tests/operations/eltwise/test_binary_uint32.py
@@ -517,10 +517,10 @@ def test_binary_comp_logical_ops_uint32_edge_cases(ttnn_op, device):
     )
 
     golden_function = ttnn.get_golden_function(ttnn_op)
-    torch_output_tensor = golden_function(torch_input_tensor_a, torch_input_tensor_b, device=device)
+    torch_output_tensor = golden_function(torch_input_tensor_a, torch_input_tensor_b, device=device).to(torch.int64)
 
     output_tensor = ttnn_op(input_tensor_a, input_tensor_b)
-    output_tensor = ttnn.to_torch(output_tensor)
+    output_tensor = ttnn.to_torch(output_tensor).to(torch.int64)
 
     assert torch.equal(output_tensor, torch_output_tensor)
